# cloud/openapi.yaml
openapi: 3.1.0
info:
  title: Jarvis Tax Assistant API
  description: |
    AI-powered tax and compliance assistant API for Indian tax professionals.
    Covers GST, Income Tax, Customs, Company Law, and FEMA.
    
    **Authentication**: Bearer JWT token or X-API-Key header.
    **License**: AGPL-3.0 (open source) / Commercial (enterprise features).
  version: 1.0.0
  contact:
    email: api@jarvis-tax.ai
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:3001
    description: Development server
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name, profession]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string }
                profession: { type: string, enum: [ca, lawyer, tax-consultant, cs, cma, other] }
                firm: { type: string }
                phone: { type: string }
      responses:
        '201':
          description: Registration successful
        '409':
          description: Email already registered
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Login successful, returns tokens
        '401':
          description: Invalid credentials
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: User profile
...
  /api/v1/subscription/current:
    get:
      tags: [Subscription]
      summary: Get current subscription
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current subscription details
  /api/v1/subscription/usage:
    get:
      tags: [Subscription]
      summary: Get usage statistics
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Usage stats for today
  /api/v1/api-keys:
    get:
      tags: [API Keys]
      summary: List API keys
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of API keys
    post:
      tags: [API Keys]
      summary: Create API key
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                permissions: { type: array, items: { type: string } }
                rateLimitPerMin: { type: integer, default: 100 }
                expiresInDays: { type: integer }
      responses:
        '201':
          description: API key created (shown only once)
...
  /api/v1/rag/stats:
    get:
      summary: RAG pipeline statistics
      tags: [RAG]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Document count, chunk count, vector DB type
  /api/v1/rag/ingest:
    post:
      summary: Ingest a document into the RAG pipeline
      tags: [RAG]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, title, content]
              properties:
                id:
                  type: string
                source:
                  type: string
                  enum: [notification, circular, case-law, act, rule, article]
                domain:
                  type: string
                title:
                  type: string
                content:
                  type: string
                citation:
                  type: string
                effectiveDate:
                  type: string
                url:
                  type: string
      responses:
        "200":
          description: Number of chunks ingested
  /api/v1/rag/search:
    get:
      summary: Search ingested documents
      tags: [RAG]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: domain
          in: query
          schema:
            type: string
        - name: topK
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Search results with relevance scores
  /api/v1/subscription/webhook/razorpay:
    post:
      summary: Razorpay payment webhook
      tags: [Subscription]
      description: Called by Razorpay on payment events. Verified via X-Razorpay-Signature header.
      parameters:
        - name: X-Razorpay-Signature
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Webhook processed
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key